adapt addpicture tests
